requirements.txt
Flask
opencv-python
pytesseract
werkzeug
numpy
mysql-connector-python
python-Levenshtein
